[thirdparty] support import brpc and gtest lib from cmake (#5)
1. support import brpc lib by cmake
2. bump brpc version to 1.8.0
3. bump gtest version to 1.14.0
4. fix compile warnings

+++ src/braft/fsm_caller.cpp
@@ -63,7 +63,7 @@ int FSMCaller::run(void* meta, bthread::TaskIterator<ApplyTask>& iter) {
         return 0;
     }
     int64_t max_committed_index = -1;
-    int64_t counter = 0;
+    size_t counter = 0;
     size_t  batch_size = FLAGS_raft_fsm_caller_commit_batch;
     for (; iter; ++iter) {
         if (iter->type == COMMITTED && counter < batch_size) {
